% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pml_db.R
\name{create_pmldb}
\alias{create_pmldb}
\title{create a pmldb instance}
\usage{
create_pmldb(dir, savedb = TRUE, invalidate = FALSE)
}
\arguments{
\item{dir}{directory containing files of interest}

\item{savedb}{save created list as a .rds file in the output dir}

\item{invalidate}{overwrite existing .rds db file if it exists}
}
\description{
create a pmldb instance
}

