% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_summaries.R
\name{get_model_summary}
\alias{get_model_summary}
\title{get the parsed yml from the model file}
\usage{
get_model_summary(mdl, dir = getwd())
}
\arguments{
\item{mdl}{model file}

\item{dir}{directory of model file}
}
\description{
get the parsed yml from the model file
}

