% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_records.R
\name{get_run_records}
\alias{get_run_records}
\title{get all runs and associated recent results folder}
\usage{
get_run_records(dir = getwd(), recursive = F, filter = NULL)
}
\arguments{
\item{dir}{directory to search}

\item{recursive}{boolean}

\item{filter}{vector of run names to filter}
}
\value{
list of dir, yml_models, model_dirs, and time of scan
}
\description{
get all runs and associated recent results folder
}

