% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_multidoc_yaml.R
\name{parse_multidoc_yaml}
\alias{parse_multidoc_yaml}
\title{parse multi-document yaml to single document}
\usage{
parse_multidoc_yaml(file, ...)
}
\arguments{
\item{filename}{}
}
\description{
parse multi-document yaml to single document
}
\details{
yaml_load is unable to parse multiline markdown documents, this
extracts each individual document to a list, so the list can be
passed to the yaml parser.
}

