% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_progress.R
\name{parse_progress}
\alias{parse_progress}
\title{parse the progress.txt file to a dataframe with values by iteration}
\usage{
parse_progress(dir, file = "progress.txt")
}
\arguments{
\item{dir}{directory of progress file}

\item{file}{file name of progress.txt}
}
\value{
data.frame with column names param, value, iteration
}
\description{
parse the progress.txt file to a dataframe with values by iteration
}

