% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pml_db.R
\name{process_files}
\alias{process_files}
\title{factory method for processing files in a directory}
\usage{
process_files()
}
\description{
factory method for processing files in a directory
}
\examples{
\dontrun{
file_processor <- process_files() # initialize
# can then access different processing functions based on file of interest
file_processor$dmp$processing_fun(dir)
file_processor$IdEta$processing_fun(dir)
}
}

